Idioms Implemented

+++ pe_l1_java_learningapp/src/com/pe/l1/underConstruction/Arith.java
@@ -5,11 +5,12 @@ package com.pe.l1.underConstruction;
  */
 public class Arith {
     public static void main(String[] args) {
-        int x = 150;
+    	//IDIOM:Item-0361: Long literals should be suffixed with L (capital)
+        long x = 150L;
         String y = "10";
-        int z = Integer.parseInt(y);
+       long z = Long.parseLong(y);
         x = (x * 10) + z;
-
+        
         byte mybytevalue = -128;
         byte mynewbytevalue = (byte) (mybytevalue/2);
 
@@ -25,6 +26,7 @@ public class Arith {
             System.out.println("x print3:" +x);
             long time = System.currentTimeMillis();
             System.gc();
+            //IDIOM:Item-0142: Inject a Clock instead of hard-coding time
             System.out.println("Took " + (System.currentTimeMillis()-time) + " ms");
             long time1 =System.nanoTime();
             System.out.println("Took " + (System.currentTimeMillis()-time1) + " ns");
